iothub-python-functions/requirements.txt
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b4
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.7.0
six==1.12.0
wheel
azure-cosmosdb-table
requests